Auto-Update bad-discords list in channel
This should auto-update the bad-discords list when importing a csv so it doesnt need to be done manually. 

Channel & Message ID need to be replaced to proper ones, Message also needs to be first posted by the BOT to be edited. 
(just let the bot post the current message and it can work with that)